修复消息并发写入序号竞争
- 移除 AppendMessage 中 MAX(turn_idx)+1 的读后写序号分配
- 复用 MySQL 自增 id 作为消息稳定排序源，事务内回填 turn_idx
- 将 messages.turn_idx 调整为 BIGINT，并在启动时按需修正表结构
- 增加同 thread 并发追加消息测试，验证 turn_idx 不重复
- 补充 OpenSpec，记录多 pod 并发写入设计与后续 migration 方向

+++ internal/store/schema.sql
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS threads (
 CREATE TABLE IF NOT EXISTS messages (
     id          BIGINT AUTO_INCREMENT PRIMARY KEY,
     thread_id   VARCHAR(128) NOT NULL,
-    turn_idx    INT NOT NULL,
+    turn_idx    BIGINT NOT NULL,
     role        VARCHAR(32) NOT NULL,
     content     MEDIUMTEXT NOT NULL,
     created_at  TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
